Fix synchronization issue when writing string columns with dictionary to ORC (#14595)
Changes in https://github.com/rapidsai/cudf/pull/14295 introduced a synchronization issue in `build_dictionaries`. After stripe_dicts are initialized on the host, we copy them to the device and then launch kernels that read the dicts (device copy). However, after these kernels we deallocate buffers that are not longer needed and clear the dicts' views to these buffers on the host. The problem is that, without synchronization after the H2D copy, the host modification can be done before the H2D copy is performed, and we run the kernels with the altered state.
This PR adds a sync point to make sure the copy is done before host-side modification.

Authors:
   - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
   - Nghia Truong (https://github.com/ttnghia)
   - Alessandro Bellina (https://github.com/abellina)
   - Bradley Dice (https://github.com/bdice)