Move creation of env.yaml outside the current directory (#14476)
Moves the creation of the `env.yaml` file in build scripts to a temporary directory instead of the current directory.

During the CI builds, a `env.yaml` file is created for use with a mamba call later in the script. This is a problem when trying to reproduce CI locally when the launched docker container tries to create this file locally as root. If the user does not have root access (or equivalent permissions), the file is not created and the process aborts.

The same error occurs when creating the `test_results` directory when the `RAPIDS_TEST_DIR` is not set. In this case the script tries to create the `test_results` in the current working directory.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - AJ Schmidt (https://github.com/ajschmidt8)

URL: https://github.com/rapidsai/cudf/pull/14476