Update pre-commit hooks (#14837)
This PR updates pre-commit hook versions and reformats the YAML so its spacing is more similar to the YAML format elsewhere in cudf and in other RAPIDS repos. Feel free to review this as two separate commits: a content change, and a format change. The changes outside of `.pre-commit-config.yaml` are from minor updates in `black`, which removed some extraneous blank lines, and a few fixes requested by `ruff`.

Also, the newer version of nbqa in this PR supports Python 3.12.

Authors:
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - Matthew Roeschke (https://github.com/mroeschke)
  - Richard (Rick) Zamora (https://github.com/rjzamora)

URL: https://github.com/rapidsai/cudf/pull/14837