fix(general, funnelshift):
To be compatible with the CUDA version of __m_funnelshift_r,
we need to use shift % 64, as the shift parameter is allowed to take in
arbitrary values > 64, which actually happens in hipDF for the operation get_mask_offset_word.
Without this modulo operation, 0 values would be returned.

From CUDA documentation:

__device__ unsigned int __funnelshift_r ( unsigned int  lo, unsigned int  hi, unsigned int  shift )
Concatenate hi : lo, shift right by *shift & 31 bits*, return the least significant 32 bits.