fix (io, parquet, reader, page preprocessing): Fixed an issue that caused chunked reads with a certain byte limit to misbehave (not split the data correctly), as page sizes of string pages were not set correctly during decoding in the shared memory state.
Therefore, the computed number of output chunks was too low (in this case,
it ended up being 1 instead of 2) which caused some unit test failures.

The underlying problem is that the result returned by gpuInitStringDescriptors
is only valid on a single thread: only this single thread should be allowed
to write its value for str_len into the shared memory state.