Change chained replace inplace test to COW test for pandas 2.2 (#15049)
`test_setitem_dataframe_series_inplace` failed with pandas 2.2 because it exhibits a chained indexing behavior that raised a `FutureWarning` in pandas 2.2 and will raise in 3.0. I refactored the test to test cudf copy on write to exhibit the 3.0 behavior, but it still seems to allow this chained indexing behavior, so xfailed it for now.

Authors:
  - Matthew Roeschke (https://github.com/mroeschke)

Approvers:
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/15049