feat/fix(udf, jitify): Decrease coupling of UDF functionality between Python interface and C++ backend: The C++ backend now parses a LLVM IR UDF from Numba and replaces its function name with the one required for a specific "jitified" kernel. It no longer requires NUMBA to chose the expected UDF function name which needs to match the one used in libhipdf's Jitified kernels.
In lack of a better method, we still need to hardcode the function name of the NUMBA UDF
to "udf_funcname_from_numba_to_be_replaced_in_libhipdf", to be able to use it as a pattern
to identify the UDF in the LLVM IR on the C++ backend side.
(Side note: there may be multiple other device functions in the LLVM IR string from NUMBA!).