fix/feat(python unit tests, test_decimal.py): Fix failing test_typecast_from_int_to_decimal:
On AMD backend, cast from negative float to unsigned integer
results in large unsigned integers (on NVIDIA they happen to be casted to 0).
These large integers then do not fit into the Decimal64Dtypes used in this test.
This is UB so we take the abs value here if an unsigned integer is used as
 from_dtype. See https://github.com/AMD-AI/hipdf/issues/106.