tmpfix(unit tests, python): expect test_timedelta_series_mod_with_cudf_scalar_zero and test_timedelta_series_mod_with_scalar_zero to pass on HIP/AMD.
Background:
Both tests eventually call the operation x % 0 in the libhipdf
backend, which is UB.
On CUDA side, the integer -1 is returned as a result which
makes the test expectedly fail, as the resulting dataframe
does not match the one from Pandas.
Pandas-compatible behavior is desired though and it is
what we observe on HIP/AMD backend.
We therefore no longer mark the test as xfail for HIP/AMD.
CAUTION: The fundamental behaviour of 'x % 0' remains undefined,
so for future compiler releases
or Pandas versions, this test may need to be changed again.