fix/war(cast_ops.hip, unary_cast):
Fix UB when casting floating point value "nan" to an integer.

Details:

We need to treat casts from floating point types to integral types specially,
as the cast behaviour from nan float to an integer is UB in C++.
This fixes the unit test test_isin_numeric in test_series.py (for the Python backend),
in which this UB in libhipdf is triggered by casting a nan floating point value to an int64.
On AMD GPUs, the result of this cast operation happens to be 0 which (by chance) matches
the value of the input data in the unit test. Therefore, the test would fail as
the casted "(long) nan == 0", although nan is clearly different from 0.