For powers of 10, replace ipow with switch (#15353)
This adds a new runtime calculation of the power-of-10 needed for applying decimal scale factors with a switch statement.  This provides the fastest way of applying the scale.  Note that the multiply and divide operations are performed within the switch itself, so that the compiler sees the full instruction to optimize assembly code gen.  See code comments for details.  

This cannot be used within fixed_point (e.g. for comparison operators and rescaling) as it introduced too much register pressure to unrelated benchmarks.  It will only be used for the decimal <--> floating conversion, so it has been moved there to be in a new header file where that code will reside (in an upcoming PR).  This is part of a larger change to change the algorithm for decimal <--> floating conversion to a more accurate one that is forthcoming soon.

Authors:
  - Paul Mattione (https://github.com/pmattione-nvidia)

Approvers:
  - Mark Harris (https://github.com/harrism)
  - Muhammad Haseeb (https://github.com/mhaseeb123)

URL: https://github.com/rapidsai/cudf/pull/15353