Add external issue label and project automation (#15945)
This PR creates two new GitHub Actions around issue and PR tracking

### `external_issue_labeler.yml`
This action automatically adds a label, currently `External`, to any issue or PR that is opened by someone that is not either an owner, member, or collaborator to the cuDF repo

### `pr_issue_status_automation.yml`
This action uses the [shared workflows](https://github.com/rapidsai/shared-workflows/tree/branch-24.08/.github/workflows) in rapdsai/shared-workflows to, on open/edit/synchronize of an open PR, to:
1. Set the PR to `in progress`
2. Set all linked issues `in progress`
3. Set the PR's sprint to the current iteration
4. Set all linked issues to the current iteration

Edit triggers on edit of the PR description, (so new linked issues will get synchronized to `in progress`). Synchronize triggers on push and rebase events - this really is to cover the "what are we working on right now" because anything we touch goes into the current sprint in the project.

Authors:
  - Ben Jarmak (https://github.com/jarmak-nv)
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/15945