Explain line profiler and how to know which functions are GPU-accelerated. (#16079)
The `cudf.pandas` docs could be more explicit about which functions are accelerated. The answer is "most of the cudf API is accelerated, and you can check the results with the profiler." I've answered this question a few times so I wanted to document it.

This closes #16074.

Authors:
  - Bradley Dice (https://github.com/bdice)
  - GALI PREM SAGAR (https://github.com/galipremsagar)
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/16079